Pull setup for ValidatesUniquenessTest out of test_helper
